# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Это сервер для управления водителями.",
        "title": "driver api",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/addDriver": {
            "post": {
                "description": "Создает запись о водителе в базе данных",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "drivers"
                ],
                "summary": "Добавить нового водителя",
                "parameters": [
                    {
                        "description": "Данные водителя",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/go-server-gin_internal_domain.Driver"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/go-server-gin_internal_domain.AddDriverResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/driver/{id}": {
            "get": {
                "description": "Получает информацию о водителе по его уникальному идентификатору",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "drivers"
                ],
                "summary": "Получить информацию о водителе",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID водителя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-server-gin_internal_domain.Driver"
                        }
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/driver/{id}/full": {
            "get": {
                "description": "Получить информацию о водителе и транспортном средстве по его уникальному идентификатору водителя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "drivers"
                ],
                "summary": "Получить информацию о водителе и транспортном средстве",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID водителя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-server-gin_internal_domain.FullDriver"
                        }
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/drivers": {
            "get": {
                "description": "Получить информацию о водителях",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "drivers"
                ],
                "summary": "Получить информацию о водителях",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/go-server-gin_internal_domain.Driver"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "go-server-gin_internal_domain.AddDriverResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "go-server-gin_internal_domain.Driver": {
            "type": "object",
            "required": [
                "name",
                "score",
                "vehicle_id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "score": {
                    "type": "integer"
                },
                "vehicle_id": {
                    "type": "integer"
                }
            }
        },
        "go-server-gin_internal_domain.FullDriver": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "score": {
                    "type": "integer"
                },
                "vehicle": {
                    "$ref": "#/definitions/go-server-gin_internal_domain.Vehicle"
                }
            }
        },
        "go-server-gin_internal_domain.Vehicle": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "model": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "vendor": {
                    "type": "string"
                }
            }
        }
    }
}